requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==5.0.6
django-appconf==1.0.6
django-celery-results==2.5.1
django-rq==2.10.2
django-video-encoding==1.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
midtransclient==1.4.2
msgpack==1.0.8
packaging==24.1
pillow==10.3.0
pip-system-certs==4.0
priority==1.3.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
Redis-Sentinel-Url==1.0.1
requests==2.32.3
rq==1.16.2
rq-dashboard==0.7.0.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
zope.interface==6.4.post2
